started with javadoc and cleanup of benchmark test
